Refactored list-page to make it clean: 1)Moved modals and list to other places
