+++ client/src/context/context.ts
@@ -1,19 +1,55 @@
 import { createContext } from "react";
+import type {
+  ExtendedTodoApiState,
+  ExtendedTodoApiActions,
+} from "./useStatesAndActions";
+import { noop } from "../utils/noop"; 
 
-import type { UseTodoApiActions, UseTodoApiState } from "./useAPIServices";
-
-export type TodoContextType = [UseTodoApiState, UseTodoApiActions];
+export type TodoContextType = [ExtendedTodoApiState, ExtendedTodoApiActions];
 
 const defaultValue: TodoContextType = [
   {
+    // UseTodoApiState
     todosData: null,
     loadingTodosData: false,
     errorTodosData: null,
     todosParams: {},
+
+    // ExtendedTodoApiState additions
+    success: noop,
+    error: noop,
+    filter: "all",
+    page: 1,
+    drawerOpen: false,
+    formData: {
+      name: "",
+      description: "",
+      dueDate: "",
+    },
+    showWarningModal: false,
+    selectedTodo: null,
+    showDeleteModal: false,
+    todoToDelete: null,
+    isEditMode: false,
   },
   {
+    // UseTodoApiActions
     fetchData: async () => {},
     setParams: () => {},
+
+    // ExtendedTodoApiActions additions
+    saveToDoData: async () => {},
+    updateToDoData: async () => {},
+    deleteToDoData: async () => {},
+
+    setFilter: () => {},
+    setPage: () => {},
+    setDrawerOpen: () => {},
+    setFormData: () => {},
+    setShowWarningModal: () => {},
+    setSelectedTodo: () => {},
+    setShowDeleteModal: () => {},
+    setTodoToDelete: () => {},
   },
 ];
