+++ client/src/hooks/index.ts
@@ -3,3 +3,4 @@ export { default as useSave } from "./useSave";
 export { default as useToast } from './useToast';
 export { default as useUpdate } from './useUpdate';
 export { default as useDelete } from './useDelete';
+export {default as useTodoHandlers} from './useTodoHandlers'
